cab_halt_alerts_pipeline/cloud-run-func/requirements.txt
functions-framework==3.*
flask
gunicorn
google-cloud-bigquery